Refactor the folder structure for customvoice api samples.
Add python sample code for voice synthesis.
