+++ CustomVoice-API-Samples/CSharp/CustomVoice-API/VoiceAPI/DTO/VoiceSynthesisDefinition.cs
