+++ CustomVoice-API-Samples/CSharp/CustomVoice-API/VoiceAPI/HttpClientExtensions.cs
@@ -1,0 +1,42 @@
+﻿// <copyright file="HttpClientExtensions.cs" company="Microsoft Corporation">
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// </copyright>
+
+namespace System.Net.Http
+{
+    using System;
+    using System.Threading.Tasks;
+    public static class HttpClientExtensions
+    {
+        public static async Task<HttpResponseMessage> PatchAsync(this HttpClient client, Uri requestUri, HttpContent iContent)
+        {
+            var method = new HttpMethod("PATCH");
+            var request = new HttpRequestMessage(method, requestUri)
+            {
+                Content = iContent
+            };
+
+            HttpResponseMessage response = new HttpResponseMessage();
+            try
+            {
+                response = await client.SendAsync(request);
+            }
+            catch (TaskCanceledException e)
+            {
+                Console.WriteLine("ERROR: " + e.ToString());
+            }
+
+            return response;
+        }
+
+        public static Task<HttpResponseMessage> PatchAsync(this HttpClient client, string requestUri, HttpContent content)
+        {
+            return client.PatchAsync(CreateUri(requestUri), content);
+        }
+
+        private static Uri CreateUri(string uri)
+        {
+            return string.IsNullOrEmpty(uri) ? null : new Uri(uri, UriKind.RelativeOrAbsolute);
+        }
+    }
+}
