+++ CustomVoice-API-Samples/CSharp/CustomVoice-API/packages.config
@@ -1,0 +1,8 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.Azure.KeyVault.Core" version="1.0.0" targetFramework="net46" />
+  <package id="Newtonsoft.Json" version="11.0.2" targetFramework="net45" />
+  <package id="System.Net.Http" version="4.3.3" targetFramework="net45" requireReinstallation="true" />
+  <package id="System.Net.Http.Formatting.Extension" version="5.2.3.0" targetFramework="net45" />
+  <package id="WindowsAzure.Storage" version="9.3.3" targetFramework="net46" />
+</packages>
