+++ CustomVoice-API-Samples/Python/voiceclient.py
@@ -1,0 +1,93 @@
+import argparse
+import json
+import ntpath
+import urllib3
+import requests
+import time
+from json import dumps, loads, JSONEncoder, JSONDecoder
+import pickle
+
+
+urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
+
+parser = argparse.ArgumentParser(description='Cris client tool to submit voice synthesis requests.')
+parser.add_argument('--voices', action="store_true", default=False, help='print voice list')
+parser.add_argument('--syntheses', action="store_true", default=False, help='print synthesis list')
+parser.add_argument('--submit', action="store_true", default=False, help='submit a synthesis request')
+parser.add_argument('--concatenateResult', action="store_true", default=False, help='If concatenate result in a single wave file')
+parser.add_argument('-file', action="store",  dest="file", help='the input text file path')
+parser.add_argument('-voiceId', action="store", dest="voiceId", help='the id of the voice which used to synthesis')
+parser.add_argument('-locale', action="store", dest="locale", help='the locale information like zh-CN/en-US')
+parser.add_argument('-key', action="store", dest="key", required=True, help='the cris subscription key, like ff1eb62d06d34767bda0207acb1da7d7 ')
+parser.add_argument('-region', action="store", dest="region", required=True, help='the region information, could be centralindia, canadacentral or uksouth')
+
+args = parser.parse_args()
+
+baseAddress = 'https://%s.cris.ai/api/texttospeech/v3.0-beta1/' % args.region
+
+
+def getSubmittedSyntheses():
+    response=requests.get(baseAddress+"voicesynthesis", headers={"Ocp-Apim-Subscription-Key":args.key}, verify=False)
+    syntheses = json.loads(response.text)
+    return syntheses
+
+def getSubmittedSynthesis(id):
+    response=requests.get(baseAddress+"voicesynthesis/"+id, headers={"Ocp-Apim-Subscription-Key":args.key}, verify=False)
+    synthesis = json.loads(response.text)
+    return synthesis
+
+def getVoices():
+    response=requests.get(baseAddress+"voicesynthesis/voices", headers={"Ocp-Apim-Subscription-Key":args.key}, verify=False)
+    voices = json.loads(response.text)
+    return voices
+
+
+def submitSynthesis():
+    filename=ntpath.basename(args.file)        
+    data={'name': 'simple test', 'description': 'desc...', 'model': args.voiceId, 'locale': args.locale}
+    if args.concatenateResult:
+        properties={'ConcatenateResult': 'true'}
+        data['properties'] = json.dumps(properties)
+    files = {'script': (filename, open(args.file, 'rb'), 'text/plain')}
+    response = requests.post(baseAddress+"voicesynthesis", data, headers={"Ocp-Apim-Subscription-Key":args.key}, files=files, verify=False)
+    if response.status_code == 202:
+        location = response.headers['Operation-Location']
+        id = location.split("/")[-1]
+        print("Submit synthesis request successful")
+        return id
+    else:
+        print("Submit synthesis request failed")
+        return 0
+
+if args.voices:
+    voices = getVoices()
+    print("There are %d voices available:" % len(voices))
+    for voice in voices:
+        print ("Name: %s, Id: %s, Locale: %s, Gender: %s, PublicVoice: %s" % (voice['name'], voice['id'], voice['locale'], voice['gender'], voice['isPublicVoice']))
+
+if args.syntheses:
+    synthese = getSubmittedSyntheses()
+    print("There are %d synthesis requests submitted:" % len(synthese))
+    for synthesis in synthese:
+        print synthesis['name']
+
+if args.submit:
+    id = submitSynthesis()
+    if (id == 0):
+        exit(1)
+
+    while(1):
+        print("\r\nChecking status")
+        synthesis=getSubmittedSynthesis(id)
+        if synthesis['status'] == "Succeeded":
+            r = requests.get(synthesis['resultsUrl'])
+            filename=id + ".zip"
+            with open(filename, 'wb') as f:  
+                f.write(r.content)
+                print("Succeeded... Result file downloaded : " + filename)
+            break
+        elif synthesis['status'] == "Running":
+            print("Running...")
+        elif synthesis['status'] == "NotStarted":
+            print("NotStarted...")
+        time.sleep(10)
