+++ CustomVoice-API-Samples/Python/zh-CN.txt
@@ -1,0 +1,50 @@
+﻿兑现的是承诺,换来的是北京的绿色。
+阶级的意志,阶级的心。
+钢铁的意志,钢铁的心。
+它针对的是在中东地区出现的威胁。
+多的是自产自销的大白菜。
+其实,纳副县长的"借"车,如果换一个比较肥的"口",是根本惹不出麻烦的。
+记者仔细查听了两个号码的自动语音服务,发现的确有所不同。
+尤其是在当前美国面临内外交困的局面下,拿出实实在在的解决办法或许比绞尽脑汁给对手抹黑更能赢得选民的心。
+舅舅的话让我难以置信。
+这是甘肃、吉林、天津等十一个省区市的总和。
+这将有利于受灾群众的安置和恢复正常生活。
+在国际缉毒组织的配合下,塞警方每年都要破获多起毒品交易,保证了国际音乐盛会的顺利进行。
+海南本次行政管理体制改革对省与市县两级政府的职责进行了科学划分,进一步强化省政府的宏观管制职责,强化市县政府推动经济社会发展的职责。
+他们跳伞下去时,穿的衣服就是一身迷彩服。
+它不是用来针对来自俄罗斯的威胁。
+众所周知,每年一度的高考,对于十年寒窗苦读的学子尤为重要。
+念的真好听,赶明儿教教我。
+近年来,为了让更多的视障人士享受信息无障碍,市残联引入了多种高科技的辅具。
+他,穿了一件松垮垮的红色衬衫。
+前门大街由于历史原因,落后的市政设施,给在这里生活和经商的人们带来了诸多不便。
+山穷水尽的罗水木,典当了仅剩的财产——一块手表,换来一点钱,买了擦鞋的工具,开始了擦鞋生涯。
+雪梨与墨尔本两地的交响乐团长期竞争,现在雪梨的乐团坦承演出造假,开幕式上的关键演出系由宿敌墨尔本的乐手捉刀,这让墨尔本当地的报纸专栏作家相当得意。
+冯静在她与老师一起署名发表的题目为《<明镜>周刊与调查性报道》的文章中写道。
+回顾过去,是为了走向未来,我们这十年之旅,是制度的探索,是香港人精神面貌的探索,也是我们价值与理想的探索。
+博客上传播,造成恶劣的社会影响。
+因此包润石的谈话,颇不寻常。
+在通常情况下,由三名法官组成的预审分庭至少需要六周时间作出相关的决定。
+扳起手指数数,咱得承认齐涛说的每一件事都是事实,绝没有浮夸的成分。
+如果长得漂亮是一种错,我已经铸成大错。
+成蕙琳是朝鲜当时最性感的电影明星。
+此间媒体注意到的是,东道国日本为八国首脑们端出了太多的美味佳肴。
+红三连在连队支部一班人的带领下突击作业。
+这大大加快了治沙速度,也提高了林木的成活率。
+锡盟强调要坚持走新型工业化道路,避免其他地区在工业化初期所走过的某些弯路,从更高的起点上谋划和推进锡盟的工业化。
+这是一场与死神比速度、与时间抢生命的紧急大空运。
+应候选人的要求,共计有一百多个投票点进行了重新计票,但均没有影响此前的统计结果。
+起初,杨丽娟的父母对女儿的这种痴迷行为多次进行劝说,但因此与女儿产生了很深的隔阂,于是父母不得不将对女儿的劝说变为支持。
+一党为主或两党的国家更易遭受军事干扰。
+中国印摩崖石刻表达的中华文化与奥运精神的结合与交融,将有着重要的现实意义。首先,今年是北京奥运举办年,同时又是我国改革开放三十周年。
+还有一种让人哭笑不得的情况。
+针对美国政府的救助行为,日本财务省次官杉本和行昨天表示,对美国援救两家房贷企业的方案表示欢迎,并称这将有助于全球金融市场的稳定。
+会议讨论了贸易、金融体系发展以及国际安全领域的一些问题。
+在一九九八年,福特表示,卡特"可能会得到比我们现在所听到的更好的一些评价"。
+所以,清洁、高效的核电成了备选。
+该书宣扬来自中国、印度等发展中国家的经济威胁,声称所有工业发达国家都正面临着所谓来自中国、印度等发展中国家的对富裕的抢夺。
+李振波的妻子盛亚玲,在观看《向死而生》时心情久久不能平静。
+党对军队的绝对领导从本质上说,是建立在这种党和军队宗旨使命高度一致基础上的"无条件"领导。
+为了抢救亲人的生命,就是再难再险我们也敢飞敢闯。
+他说,粮价上涨是多种因素综合作用的结果,把全球粮食需求增加归咎于发展中国家的发展,既不符合事实,也不是解决问题的建设性态度。
+昨晚的比赛,国安队基本上占据了优势。
