--- CustomVoice-API/CustomVoice-API/VoiceAPI/DTO/DatasetIdentity.cs
@@ -1,18 +1,0 @@
-﻿// <copyright file="DatasetIdentity.cs" company="Microsoft Corporation">
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// </copyright>
-
-namespace Microsoft.SpeechServices.Cris.Http
-{
-    using System;
-
-    public sealed class DatasetIdentity
-    {
-        public DatasetIdentity(Guid id)
-        {
-            this.Id = id;
-        }
-
-        public Guid Id { get; private set; }
-    }
-}
