--- CustomVoice-API/CustomVoice-API/VoiceAPI/DTO/Synthesis.cs
@@ -1,54 +1,0 @@
-﻿// <copyright file="Synthesis.cs" company="Microsoft Corporation">
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// </copyright>
-
-namespace Microsoft.SpeechServices.Cris.Http
-{
-    using Newtonsoft.Json;
-    using System;
-    using System.Collections.Generic;
-
-    public sealed class Synthesis
-    {
-        [JsonConstructor]
-        private Synthesis(Guid id, string name, string description, string locale, DateTime createdDateTime, DateTime lastActionDateTime, string status, Uri resultsUrl, IReadOnlyDictionary<string, string> properties)
-        {
-            this.Id = id;
-            this.Name = name;
-            this.Description = description;
-            this.CreatedDateTime = createdDateTime;
-            this.LastActionDateTime = lastActionDateTime;
-            this.Status = status;
-            this.Locale = locale;
-            this.Properties = properties;
-            this.ResultsUrl = resultsUrl;
-        }
-
-        /// <inheritdoc />
-        public string Name { get; set; }
-
-        /// <inheritdoc />
-        public string Description { get; set; }
-
-        /// <inheritdoc />
-        public string Locale { get; set; }
-
-        /// <inheritdoc />
-        public Uri ResultsUrl { get; set; }
-
-        public Guid Id { get; set; }
-
-        /// <inheritdoc />
-        public DateTime CreatedDateTime { get; set; }
-
-        /// <inheritdoc />
-        public DateTime LastActionDateTime { get; set; }
-
-        /// <inheritdoc />
-        public string Status { get; set; }
-
-        public string StatusMessage { get; set; }
-
-        public IReadOnlyDictionary<string, string> Properties { get; set; }
-    }
-}
